extract ConfigureIcsAction
