+++ plugins/settings-repository/src/actions/ConfigureIcsAction.kt
@@ -1,0 +1,48 @@
+// Copyright 2000-2018 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+package org.jetbrains.settingsRepository.actions
+
+import com.intellij.configurationStore.StateStorageManagerImpl
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.components.stateStore
+import com.intellij.openapi.fileChooser.FileChooserDescriptorFactory
+import com.intellij.openapi.project.DumbAwareAction
+import com.intellij.openapi.ui.TextFieldWithBrowseButton
+import com.intellij.ui.components.dialog
+import com.intellij.ui.layout.*
+import org.jetbrains.settingsRepository.createMergeActions
+import org.jetbrains.settingsRepository.icsManager
+import org.jetbrains.settingsRepository.icsMessage
+import kotlin.properties.Delegates
+
+internal class ConfigureIcsAction : DumbAwareAction() {
+  override fun actionPerformed(e: AnActionEvent) {
+    var urlTextField: TextFieldWithBrowseButton by Delegates.notNull()
+    icsManager.runInAutoCommitDisabledMode {
+      val panel = panel {
+        row(icsMessage("settings.upstream.url")) {
+          urlTextField = textFieldWithBrowseButton(value = icsManager.repositoryManager.getUpstream(),
+                                                   browseDialogTitle = "Choose Local Git Repository",
+                                                   fileChooserDescriptor = FileChooserDescriptorFactory.createSingleFolderDescriptor())
+        }
+      }
+      dialog(title = icsMessage("settings.panel.title"),
+             panel = panel,
+             focusedComponent = urlTextField,
+             project = e.project,
+             createActions = {
+                                          createMergeActions(e.project, urlTextField, it)
+                                        }).show()
+    }
+  }
+
+  override fun update(e: AnActionEvent) {
+    val application = ApplicationManager.getApplication()
+    if (application.isUnitTestMode) {
+      return
+    }
+
+    e.presentation.isEnabledAndVisible = icsManager.isActive || !(application.stateStore.storageManager as StateStorageManagerImpl).compoundStreamProvider.enabled
+    e.presentation.icon = null
+  }
+}
